src/helper_functions.py
import librosa
import numpy as np
import cv2

def preprocess_audio(audio_path):
    """
    Preprocess audio file for model input.

    Args:
        audio_path (str): Path to the audio file.

    Returns:
        np.ndarray: Preprocessed audio data.
    """
    # Load and preprocess audio file
    audio, sr = librosa.load(audio_path, sr=None)
    # Further preprocessing as required
    return audio

def preprocess_video(video_path):
    """
    Preprocess video file for model input.

    Args:
        video_path (str): Path to the video file.

    Returns:
        np.ndarray: Preprocessed video frames.
    """
    video_frames = []
    cap = cv2.VideoCapture(video_path)

    while cap.isOpened():
        ret, frame = cap.read()
        if not ret:
            break
        # Preprocess the frame (e.g., resizing, normalization)
        frame = cv2.resize(frame, (224, 224))  # Example resize
        frame = frame / 255.0  # Example normalization
        video_frames.append(frame)

    cap.release()
    return np.array(video_frames)

def preprocess_animation(animation_path):
    """
    Preprocess animation video for model input.

    Args:
        animation_path (str): Path to the animation video file.

    Returns:
        np.ndarray: Preprocessed animation frames.
    """
    return preprocess_video(animation_path)
















































